+++ src/main/java/roomescape/auth/jwt/JwtTokenProvider.java
@@ -5,7 +5,7 @@ import io.jsonwebtoken.security.Keys;
 import java.nio.charset.StandardCharsets;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import roomescape.member.Member;
+import roomescape.member.domain.Member;
 
 @Component
 public class JwtTokenProvider {
