+++ src/main/java/roomescape/member/repository/MemberRepository.java
@@ -1,8 +1,9 @@
-package roomescape.member;
+package roomescape.member.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import roomescape.member.domain.Member;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
