+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.theme;
+package roomescape.theme.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -10,6 +10,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.net.URI;
 import java.util.List;
+import roomescape.theme.repository.ThemeRepository;
+import roomescape.theme.domain.Theme;
 
 @RestController
 public class ThemeController {
