+++ src/main/java/roomescape/theme/repository/ThemeRepository.java
@@ -1,7 +1,8 @@
-package roomescape.theme;
+package roomescape.theme.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import roomescape.theme.domain.Theme;
 
 @Repository
 public interface ThemeRepository extends JpaRepository<Theme, Long> {
