+++ src/main/java/roomescape/time/TimeService.java
@@ -5,8 +5,8 @@ import roomescape.reservation.domain.Reservation;
 
 import java.util.List;
 import roomescape.reservation.repository.ReservationRepository;
-import roomescape.theme.Theme;
-import roomescape.theme.ThemeRepository;
+import roomescape.theme.domain.Theme;
+import roomescape.theme.repository.ThemeRepository;
 
 @Service
 public class TimeService {
