+++ src/main/java/roomescape/waiting/controller/WaitingController.java
@@ -1,4 +1,4 @@
-package roomescape.waiting;
+package roomescape.waiting.controller;
 
 import java.net.URI;
 import org.springframework.http.ResponseEntity;
@@ -8,6 +8,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.domain.LoginMember;
+import roomescape.waiting.dto.WaitingRequest;
+import roomescape.waiting.dto.WaitingResponse;
+import roomescape.waiting.service.WaitingService;
 
 @RestController
 public class WaitingController {
